requirements.txt
torch 
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cpu
keras==2.11.0
tensorflow==2.11.0

git+https://github.com/openai/CLIP.git
git+https://github.com/huggingface/transformers.git

pillow==11.0.0
tqdm==4.67.1
scikit-learn==1.6.1
pymongo==4.12.0
python-dotenv==1.1.0
motor==3.7.0
einops==0.8.1
numpy==1.26.4
requests==2.32.3
httpx==0.28.1
googletrans==4.0.2
starlette==0.46.1
authlib==1.5.2
apscheduler==3.11.0
cryptography==44.0.2
uvicorn[standard]
fastapi[all]
google-generativeai
protobuf==3.19.6